+++ cobigen/cobigen-core/src/main/java/com/devonfw/cobigen/impl/util/ConfigurationFinder.java
@@ -23,6 +23,7 @@ import com.devonfw.cobigen.api.exception.InvalidConfigurationException;
 import com.devonfw.cobigen.api.util.CobiGenPaths;
 import com.devonfw.cobigen.api.util.TemplatesJarUtil;
 import com.devonfw.cobigen.api.util.mavencoordinate.MavenCoordinate;
+import com.devonfw.cobigen.api.util.mavencoordinate.MavenCoordinateState;
 import com.devonfw.cobigen.impl.config.ConfigurationProperties;
 
 /**
@@ -89,8 +90,8 @@ public class ConfigurationFinder {
       mavenCoordinates = Arrays.asList(props.getProperty(templateSetsInstalled).split(","));
     }
 
-    List<MavenCoordinate> hiddenIds = MavenCoordinateUtil.convertToMavenCoordinates(hiddenIdsString);
-    List<MavenCoordinate> convertedMavenCoordinates = MavenCoordinateUtil.convertToMavenCoordinates(mavenCoordinates);
+    List<MavenCoordinate> hiddenIds = MavenCoordinateState.convertToMavenCoordinates(hiddenIdsString);
+    List<MavenCoordinate> convertedMavenCoordinates = MavenCoordinateState.convertToMavenCoordinates(mavenCoordinates);
 
     ConfigurationProperties configurationProperties = new ConfigurationProperties(groupIds, useSnapshots, hiddenIds,
         convertedMavenCoordinates);
